# Swagger/Cargohub V2/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "CargoHub V2",
      "version": "v2"
    },
    "paths": {
...
      "/api/v2/Inventories": {
        "get": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "page",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 1
              }
            },
            {
              "name": "pageSize",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "integer",
                "format": "int32",
                "default": 10
              }
            },
            {
              "name": "sortOrder",
              "in": "query",
              "style": "form",
              "schema": {
                "type": "string",
                "default": "asc"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "post": {
          "tags": [
            "Inventories"
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Inventories/{id}": {
        "get": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "put": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/Inventory"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "patch": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": { }
              },
              "text/json": {
                "schema": { }
              },
              "application/*+json": {
                "schema": { }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        },
        "delete": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
      "/api/v2/Inventories/{id}/force": {
        "delete": {
          "tags": [
            "Inventories"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "style": "simple",
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Inventory": {
          "type": "object",
          "properties": {
            "Id": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "Item_Id": {
              "type": "string",
              "nullable": true
            },
            "Description": {
              "type": "string",
              "nullable": true
            },
            "Item_Reference": {
              "type": "string",
              "nullable": true
            },
            "Locations": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              },
              "nullable": true
            },
            "Total_On_Hand": {
              "type": "integer",
              "format": "int32"
            },
            "Total_Expected": {
              "type": "integer",
              "format": "int32"
            },
            "Total_Ordered": {
              "type": "integer",
              "format": "int32"
            },
            "Total_Allocated": {
              "type": "integer",
              "format": "int32"
            },
            "Total_Available": {
              "type": "integer",
              "format": "int32"
            },
            "Created_At": {
              "type": "string",
              "nullable": true
            },
            "Updated_At": {
              "type": "string",
              "nullable": true
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }